feat: New Fm Options
- add front-matter option escapeBraces
- add front-matter option moduleScript

This commit caters a bit more to the use of the Svelte component behind a markdown document.  The front-matter option `escapeBraces` can escape all braces in the document, and `moduleScript` allows for adding additional script to the document's `<script module>`.  Overall usage of Svelte syntax features in markdown remains limited, however.